Add files via upload
There was a small bug where adding a separate JS file in the <head> would execute the code and not actually find the body tag since it had not been created by the page yet. So now console elements are created under the HTML tag, which should be fine for just debugging.